chore: move pages that need auth under the (auth)
