+++ src/app/(auth)/(protected)/(main)/layout.jsx
@@ -3,7 +3,7 @@ import { Outlet } from "react-router-dom";
 import { QueryClient } from "@tanstack/query-core";
 import { QueryClientProvider } from "@tanstack/react-query";
 
-import NavButtons from "../../../components/NavButtons";
+import NavButtons from "../../../../components/NavButtons";
 
 import styles from "./layout.module.css";
