+++ src/app/(auth)/(protected)/(main)/layout.module.css
