+++ src/app/(auth)/(protected)/(main)/photos/page.module.css
