+++ src/app/(auth)/(protected)/(main)/photos/usePhotos.js
@@ -1,5 +1,5 @@
 import { useInfiniteQuery } from "@tanstack/react-query";
-import supabase from "../../../../supabase-client/supabase";
+import supabase from "../../../../../supabase-client/supabase";
 
 function transform(data) {
   const transformed = data.map((p) => ({
