+++ src/app/page.jsx
@@ -2,15 +2,10 @@ import React, { useEffect } from "react";
 import { Button } from "antd-mobile";
 import { Navigate, useNavigate } from "react-router-dom";
 
-import { useUser } from "../contexts/AuthContext";
 import LoadingPage from "../components/LoadingPage";
 
 function Page() {
   const navigate = useNavigate();
-  const [user, loading] = useUser();
-
-  if (loading) return <LoadingPage />;
-  if (user) return <Navigate to="/home" />;
 
   return (
     <div>
